# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FisioGest Pro API - Clinical management platform for physiotherapy, aesthetics and pilates
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientListResponse"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientRequest"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientRequest"
      responses:
        "200":
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete a patient
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Patient deleted
  /procedures:
    get:
      operationId: listProcedures
      tags: [procedures]
      summary: List all procedures
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [fisioterapia, estetica, pilates]
      responses:
        "200":
          description: List of procedures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Procedure"
    post:
      operationId: createProcedure
      tags: [procedures]
      summary: Create a procedure
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcedureRequest"
      responses:
        "201":
          description: Procedure created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Procedure"
  /procedures/{id}:
    put:
      operationId: updateProcedure
      tags: [procedures]
      summary: Update a procedure
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcedureRequest"
      responses:
        "200":
          description: Procedure updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Procedure"
    delete:
      operationId: deleteProcedure
      tags: [procedures]
      summary: Delete a procedure
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /reports/monthly-revenue:
    get:
      operationId: getMonthlyRevenue
      tags: [reports]
      summary: Monthly revenue report
      security:
        - BearerAuth: []
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly revenue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenue"
  /reports/procedure-revenue:
    get:
      operationId: getProcedureRevenue
      tags: [reports]
      summary: Revenue by procedure
      security:
        - BearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Revenue by procedure
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcedureRevenue"
  /reports/schedule-occupation:
    get:
      operationId: getScheduleOccupation
      tags: [reports]
      summary: Schedule occupation report
      security:
        - BearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Schedule occupation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleOccupation"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cpf:
          type: string
        birthDate:
          type: string
          format: date
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        profession:
          type: string
        emergencyContact:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - cpf
        - phone
        - createdAt
    PatientDetail:
      allOf:
        - $ref: "#/components/schemas/Patient"
        - type: object
          properties:
            totalAppointments:
              type: integer
            lastAppointment:
              type: string
              format: date-time
            totalSpent:
              type: number
    PatientListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreatePatientRequest:
      type: object
      properties:
        name:
          type: string
        cpf:
          type: string
        birthDate:
          type: string
          format: date
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        profession:
          type: string
        emergencyContact:
          type: string
        notes:
          type: string
      required:
        - name
        - cpf
        - phone
    UpdatePatientRequest:
      type: object
      properties:
        name:
          type: string
        cpf:
          type: string
        birthDate:
          type: string
          format: date
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        profession:
          type: string
        emergencyContact:
          type: string
        notes:
          type: string
    Procedure:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [fisioterapia, estetica, pilates]
        durationMinutes:
          type: integer
        price:
          type: number
        cost:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - durationMinutes
        - price
        - createdAt
    CreateProcedureRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [fisioterapia, estetica, pilates]
        durationMinutes:
          type: integer
        price:
          type: number
        cost:
          type: number
        description:
          type: string
      required:
        - name
        - category
        - durationMinutes
        - price
...
    MonthlyRevenue:
      type: object
      properties:
        month:
          type: integer
        monthName:
          type: string
        revenue:
          type: number
        expenses:
          type: number
        profit:
          type: number
      required:
        - month
        - monthName
        - revenue
        - expenses
        - profit
    ProcedureRevenue:
      type: object
      properties:
        procedureId:
          type: integer
        procedureName:
          type: string
        category:
          type: string
        totalRevenue:
          type: number
        totalSessions:
          type: integer
        averageTicket:
          type: number
      required:
        - procedureId
        - procedureName
        - category
        - totalRevenue
        - totalSessions
        - averageTicket
    ScheduleOccupation:
      type: object
      properties:
        totalSlots:
          type: integer
        occupiedSlots:
          type: integer
        occupationRate:
          type: number
        canceledCount:
          type: integer
        noShowCount:
          type: integer
        byDayOfWeek:
          type: array
          items:
            type: object
            properties:
              dayOfWeek:
                type: string
              count:
                type: integer
            required:
              - dayOfWeek
              - count
      required:
        - totalSlots
        - occupiedSlots
        - occupationRate
        - canceledCount
        - noShowCount
        - byDayOfWeek